+++ client/presentation/src/main/java/com/andone/memorip/presentation/screen/placelist/PlaceListViewModel.kt
@@ -9,9 +9,9 @@ import com.andone.memorip.presentation.model.toUiModel
 import com.andone.memorip.presentation.screen.placelist.model.PlaceListAction
 import com.andone.memorip.presentation.screen.placelist.model.PlaceListEvent
 import com.andone.memorip.presentation.screen.placelist.model.PlaceListUiState
-import com.andone.memorip.presentation.placelist.model.RegionUiModel
-import com.andone.memorip.presentation.placelist.model.SelectedRegionState
-import com.andone.memorip.presentation.placelist.model.toUiModel
+import com.andone.memorip.presentation.screen.placelist.model.RegionUiModel
+import com.andone.memorip.presentation.screen.placelist.model.SelectedRegionState
+import com.andone.memorip.presentation.screen.placelist.model.toUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.Channel.Factory.BUFFERED
