+++ client/presentation/src/main/java/com/andone/memorip/presentation/screen/placelist/model/SelectedRegionState.kt
@@ -1,4 +1,4 @@
-package com.andone.memorip.presentation.placelist.model
+package com.andone.memorip.presentation.screen.placelist.model
 
 data class SelectedRegionState(
     val parents: List<RegionUiModel> = emptyList(),
